videostorage/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
decorator==4.4.2
Django==4.0
django-filter==22.1
djangorestframework==3.13.1
Flask==2.1.3
idna==3.3
imageio==2.19.5
imageio-ffmpeg==0.4.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
moviepy==1.0.3
numpy==1.23.1
Pillow==9.2.0
proglog==0.1.10
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
tqdm==4.64.0
tzdata==2022.1
urllib3==1.26.10
Werkzeug==2.1.2
